Create a file named data.js that contains the data to be displayed.
Import the array from data.js and create a state to get the item in the array
accordidng to a certain index.

Create a default propTypes to warn that a prop named data is required to be
sent as the component Reviews is created.

Install and use react icons from react-icons library and display the icons
inside buttons.